backend-fastapi/ml/flashcard_interval_model/requirements.txt
google-cloud-bigquery>=3.13.0
pandas>=2.1.0
db-dtypes>=1.2.0
scikit-learn==1.3.2
joblib>=1.3.0
numpy>=1.24.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
kfp>=2.0.0
